requirements.txt
Flask==3.0.0
pymongo==4.6.1
dnspython==2.4.2
requests==2.31.0
nmap==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.2.1
numpy==1.26.4
nltk==3.8.1
scikit-learn==1.4.1.post1
flask-wtf==1.2.1
reportlab==4.0.7  # For generating PDF reports
